weather-etl-extract/requirements.txt
functions-framework==3.8.0
google-cloud-storage==2.18.0
google-cloud-bigquery==3.25.0
requests==2.32.3
